+++ app/controllers/application_controller.rb
@@ -3,10 +3,25 @@ class ApplicationController < ActionController::Base
   include Authorization
   include BulletHelper
   include Geolocation
-  include TimezoneSetter
 
   protect_from_forgery with: :exception
 
+  around_action :timezone=
+
   responders :flash
   respond_to :html
+
+  private
+
+  def timezone=(&block)
+    Time.use_zone(timezone, &block)
+  end
+
+  def timezone
+    current_user&.timezone || geolocation_timezone
+  end
+
+  def geolocation_timezone
+    Timezone.lookup(*current_coordinates).name
+  end
 end
